Tidied up the file structure by moving the calculator classes to their own package.
