+++ src/main/java/ntnu/gruppe21/calculators/PurchaseCalculator.java
@@ -1,4 +1,7 @@
-package ntnu.gruppe21;
+package ntnu.gruppe21.calculators;
+
+import ntnu.gruppe21.Share;
+import ntnu.gruppe21.TransactionCalculator;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
